fix for outbound buffer
