+++ IoT/IoT_embedded/lib/drivers/display.c
@@ -1,5 +1,6 @@
 #include "display.h"
 #include "includes.h"
+#include <stdbool.h>
 //LATCH
 #define LATCH_BIT PG5
 #define LATCH_DDR DDRG
@@ -198,4 +199,24 @@ void display_dead(){
 
 void display_empty(){
     display_setValues(17, 17, 17, 17);
+}
+
+static bool first = false;
+
+void displayAndDie() {
+    if (!first){
+        display_empty();
+        disable_timer_b();
+    }
+    first = false;
+}
+
+void asyncDisableDisplayAfterMs(int ms){
+    first = true;
+    void (*disablePointer)(void) = &displayAndDie;
+    periodic_task_init_b(disablePointer, ms);
+}
+
+void init_display() {
+    display_init();
 }
