+++ IoT/IoT_embedded/lib/drivers/leds.h
@@ -7,3 +7,7 @@ void leds_turnOn(uint8_t led_no);//takes in the value 1, 2, 3 or 4, each number 
 void leds_turnOff(uint8_t led_no);//takes in the value 1, 2, 3 or 4, each number corresponding to a LED. 
 
 void leds_toggle(uint8_t led_no);//takes in the value 1, 2, 3 or 4, each number corresponding to a LED. 
+
+
+void turnOffAll(void);
+void ledAnimation(void);
